backend/requirements.txt
wheel
gunicorn
Werkzeug==2.2.2
Flask==2.2.5
flask-sqlalchemy
flask-login
flask-jwt-extended
flask-cors
flask-migrate
psycopg2-binary
requests
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-storage
google-cloud-secret-manager
google-cloud-logging
DateTime
beautifulsoup4
chromadb>=0.4.6
langchain>=0.0.267
langchain-openai>=0.0.2
langchain_core
langgraph
langchain-groq==0.3.2
openai>=1.3.0
pandas>=2.0.3
numpy>=1.24.3
pydantic>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
nltk>=3.8.1
spacy>=3.6.0
mlflow
scikit-learn
dataclasses
typing
